dataVisualization/requirements.txt
about-time==4.2.1
alive-progress==3.1.5
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
Faker==24.11.0
grapheme==0.6.0
greenlet==3.0.3
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nltk==3.8.1
packaging==24.0
parso==0.8.4
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
regex==2024.4.16
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
pyecharts==2.0.5
